add - fzf actions and statusline change
the fzf_actions should be default so might not be necessary. In any case
a ctrl-p followed by a ctrl-v will open the file in a split.

also, change the status line to include the current directory as well as
the current filename. this is handy in ember-cli projects that use the
--pod flag since it will result in a lot of files with the same name.